+++ adventure/package.json
@@ -15,6 +15,7 @@
     "@parcel/packager-ts": "2.7.0",
     "@parcel/transformer-typescript-types": "2.7.0",
     "@types/node": "16",
+    "nearley": "^2.20.1",
     "parcel": "^2.7.0",
     "ts-node": "^10.9.1",
     "typescript": "~4.7.3"
