+++ adventure/src/engine/state/testCondition.ts
@@ -1,6 +1,6 @@
-import { StateCondition } from "../dsl/ast-types";
-import { GameStateManager } from "./engine-types";
-import { GameWorld } from "../dsl/world-types";
+import { StateCondition } from "../../dsl/ast-types";
+import { GameStateManager } from "./types";
+import { GameWorld } from "../../dsl/world-types";
 
 export const testCondition = <Game extends GameWorld>(
   condition: StateCondition<Game>,
